Adição de qrys e fixDelete da arvore

+++ t4/src/Utils/TabelaHash/tabelaHash.h
@@ -13,7 +13,7 @@ TabelaHash criarTabelaHash(int tam, bool(*equals)(Objeto objeto, char* chave),
 void destruirTabelaHash(TabelaHash t);
 void TabelaHash_inserir(TabelaHash t, Objeto objeto);
 Objeto TabelaHash_getObjeto(TabelaHash t, char* chave);
-int TabelaHash_getChaveUnica(TabelaHash t, char* chave);
+long long TabelaHash_getChaveUnica(TabelaHash t, char* chave);
 bool TabelaHash_removerObjeto(TabelaHash t, char* chave);
 
 #endif
